rgw/op: kill compiling warning
rgw/rgw_op.cc: In member function 'virtual int RGWCreateBucket::verify_permission()':
rgw/rgw_op.cc:1810:37: warning: comparison between signed and unsigned integer expressions [-Wsign-compare]
     if (buckets.count() >= s->user->max_buckets) {

+++ src/rgw/rgw_op.cc
@@ -1807,7 +1807,7 @@ int RGWCreateBucket::verify_permission()
     if (op_ret < 0)
       return op_ret;
 
-    if (buckets.count() >= s->user->max_buckets) {
+    if ((int)buckets.count() >= s->user->max_buckets) {
       return -ERR_TOO_MANY_BUCKETS;
     }
   }
